backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
simplekml
Pillow
shapely